+++ opencascade-hs/cpp/hs_RWGltf_CafWriter.cpp
@@ -1,5 +1,6 @@
 #include <RWGltf_CafWriter.hxx>
-#include <TColStd_IndexedDataMapOfStringString.hxx>
+#include <NCollection_IndexedDataMap.hxx>
+#include <TCollection_AsciiString.hxx>
 #include "hs_Exception.h"
 #include "hs_RWGltf_CafWriter.h"
 
@@ -12,7 +13,7 @@ void hs_delete_RWGltf_CafWriter(RWGltf_CafWriter * theWriter){
 }
 
 void hs_RWGltf_CafWriter_Perform(
-    RWGltf_CafWriter * theWriter, Handle(TDocStd_Document) * theDocument, TColStd_IndexedDataMapOfStringString * theFileInfo, Message_ProgressRange * theProgress,
+    RWGltf_CafWriter * theWriter, Handle(TDocStd_Document) * theDocument, INDEXED_DATA_MAP(TCollection_AsciiString, TCollection_AsciiString) * theFileInfo, Message_ProgressRange * theProgress,
     HSExceptionType* exType, void ** exPtr
 ){
     hs_handleExVoid(