+++ opencascade-hs/cpp/hs_RWObj_CafWriter.cpp
@@ -1,5 +1,6 @@
 #include <RWObj_CafWriter.hxx>
-#include <TColStd_IndexedDataMapOfStringString.hxx>
+#include <NCollection_IndexedDataMap.hxx>
+#include <TCollection_AsciiString.hxx>
 #include "hs_Exception.h"
 #include "hs_RWObj_CafWriter.h"
 
@@ -12,7 +13,7 @@ void hs_delete_RWObj_CafWriter(RWObj_CafWriter * writer){
 }
 
 void hs_RWObj_CafWriter_Perform(
-    RWObj_CafWriter * theWriter, Handle(TDocStd_Document) * theDocument, TColStd_IndexedDataMapOfStringString * theFileInfo, Message_ProgressRange * theProgress,
+    RWObj_CafWriter * theWriter, Handle(TDocStd_Document) * theDocument, INDEXED_DATA_MAP(TCollection_AsciiString, TCollection_AsciiString) * theFileInfo, Message_ProgressRange * theProgress,
     HSExceptionType* exType, void ** exPtr
 ){
     hs_handleExVoid(