+++ opencascade-hs/src/OpenCascade/TCollection.hs
@@ -1,0 +1,4 @@
+module OpenCascade.TCollection
+( module OpenCascade.TCollection.Types
+) where
+import OpenCascade.TCollection.Types