+++ opencascade-hs/src/OpenCascade/TCollection/Types.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.TCollection.Types
+( AsciiString
+) where
+
+data AsciiString